Add .arch_extension directive to compile ARM smc instruction
[Defect / Enhancement / New Feature]
Defect
[Problem]
smc instructions do not compile with newer versions
of GCC toolchain.

[Solution]
Add .arch_extension sec inline to enable
use of the smc instruction.

Also fix a compiler error with the new toolchain in
audio_vdriver.c code.
[Reviewers]
Shankar B , Sandeep Patil , Vikram
http://rb.sj.broadcom.com/r/50619/
<<BYPASS_CODE_REVIEW>>
[Testing]
    [Test Cases]
		Compile kernel & u-boot
		after the changes
    [Label/CL# Synced To]
	HEAD on sdb-rhea-android-jb
    [Platform]
	Android-Rhea
[BROADCOM INTERNAL NOTES]


test

+++ sound/caph/audio_driver/audio_vdriver.c
@@ -1330,11 +1330,11 @@ void AUDDRV_SetAudioMode_Speaker(SetAudioMode_Sp_t param)
 						/*mono output*/
 
 					if (path->srcmRoute[i][j].inChnl ==
-							CAPH_SRCM_CH5
+							 (int )	CAPH_SRCM_CH5
 						|| path->srcmRoute[i][j].inChnl
-							== CAPH_SRCM_PASSCH1
+							== (int) CAPH_SRCM_PASSCH1
 						|| path->srcmRoute[i][j].inChnl
-							== CAPH_SRCM_PASSCH2) {
+							==( int ) CAPH_SRCM_PASSCH2) {
 						/*only on stereo inputs.*/
 						if (mixInGain != GAIN_NA &&
 						mixInGainR != GAIN_NA) {
